cmd: return a user friendly message if the user is not authenticated
This PR adds a pre-hook to relevant subcommands that checks if the user
is authenticated already. If not, it returns a user friendly error
message and tells them to authenticate via `pscale auth login`.

Before:

```
$ pscale db list
Error: required flag(s) "org" not set
```

After:

```
$ pscale db list
Error: not authenticated yet. Please run 'pscale auth login'
```

We didn't add the pre-hook to `rootCmd`, because that would block the
user executing the `pscale auth` commands.

closes https://github.com/planetscale/project-big-bang/issues/205
